+++ src/core/i18n/common/fr.js
@@ -47,4 +47,5 @@ registerTranslations('fr', {
   manage_label: 'Gérer',
   accounts_label: 'Comptes',
   goto_manage: 'Gérer →',
+  account_label: 'Compte',
 });